+++ addOns/domxss/CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Update minimum ZAP version to 2.11.1.
 - Use Network add-on to proxy browser requests.
 
+### Fixed
+- Stop the proxy when ZAP shuts down.
+
 ## [12] - 2021-12-06
 ### Changed
 - Dependency updates.